Lectures/08 Level III/05 FAQ LLM Telegram Bot/requirements.txt
emoji==2.13.0
loguru==0.7.0
openai==1.46.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
